tests/requirements-test.txt
pyinstaller
setuptools
packaging
wheel
pytest
pytest-mock
